refactor: LoginRequest/Response 위치 global로 변경
