+++ src/main/java/com/dash/leap/domain/user/service/UserService.java
@@ -6,10 +6,10 @@ import com.dash.leap.domain.community.repository.PostRepository;
 import com.dash.leap.domain.mission.entity.enums.MissionStatus;
 import com.dash.leap.domain.mission.repository.MissionRecordRepository;
 import com.dash.leap.domain.user.dto.request.ChatbotSettingRequest;
-import com.dash.leap.domain.user.dto.request.LoginRequest;
+import com.dash.leap.global.auth.dto.request.LoginRequest;
 import com.dash.leap.domain.user.dto.request.UserRegisterRequest;
 import com.dash.leap.domain.user.dto.response.ChatbotSettingResponse;
-import com.dash.leap.domain.user.dto.response.LoginResponse;
+import com.dash.leap.global.auth.dto.response.LoginResponse;
 import com.dash.leap.domain.user.dto.response.MyPageResponse;
 import com.dash.leap.domain.user.dto.response.UserRegisterResponse;
 import com.dash.leap.domain.user.entity.User;
