+++ src/main/java/com/dash/leap/global/auth/dto/request/LoginRequest.java
@@ -1,4 +1,4 @@
-package com.dash.leap.domain.user.dto.request;
+package com.dash.leap.global.auth.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
