+++ src/main/java/com/dash/leap/global/auth/dto/response/LoginResponse.java
@@ -1,4 +1,4 @@
-package com.dash.leap.domain.user.dto.response;
+package com.dash.leap.global.auth.dto.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
